+++ Documentation/doxygen-html/dir_0dffaf26f41c2932de490572fa27ee8a.js
@@ -1,5 +1,6 @@
 var dir_0dffaf26f41c2932de490572fa27ee8a =
 [
     [ "AnaCanScan.cpp", "AnaCanScan_8cpp.html", "AnaCanScan_8cpp" ],
-    [ "AnaCanScan.h", "AnaCanScan_8h.html", "AnaCanScan_8h" ]
+    [ "AnaCanScan.h", "AnaCanScan_8h.html", "AnaCanScan_8h" ],
+    [ "AnaGateErrors.h", "AnaGateErrors_8h.html", "AnaGateErrors_8h" ]
 ];