+++ Documentation/doxygen-html/search/all_13.js
@@ -1,4 +1,5 @@
 var searchData=
 [
-  ['version_2eh_301',['VERSION.h',['../VERSION_8h.html',1,'']]]
+  ['vendorbaudrate2userbaudrate_317',['vendorBaudRate2UserBaudRate',['../classSTCanScan.html#aa57c48594c8a70a0331f07fb4ad208fd',1,'STCanScan']]],
+  ['version_2eh_318',['VERSION.h',['../VERSION_8h.html',1,'']]]
 ];