+++ Documentation/doxygen-html/search/all_4.js
@@ -1,13 +1,13 @@
 var searchData=
 [
-  ['delete_5fmaps_72',['delete_maps',['../classCanModule_1_1Diag.html#aa79c20827b6e33fa35d4106d9aa479b8',1,'CanModule::Diag']]],
-  ['deletecanhandleofportip_73',['deleteCanHandleOfPortIp',['../classAnaCanScan.html#a1b62497ac5f0edcf9895c9f8cfad6b50',1,'AnaCanScan']]],
-  ['deviceid_74',['deviceID',['../structudevanalyserforpeak__ns_1_1UdevAnalyserForPeak_1_1PEAK__device__t.html#ac29ee824c7e4d8441a815af4c0523f6a',1,'udevanalyserforpeak_ns::UdevAnalyserForPeak::PEAK_device_t']]],
-  ['diag_75',['Diag',['../classCanModule_1_1Diag.html',1,'CanModule::Diag'],['../classCanModule_1_1Diag.html#a282a54b7181d085ba0abccca8fcb0874',1,'CanModule::Diag::Diag()']]],
-  ['diag_2ecpp_76',['Diag.cpp',['../Diag_8cpp.html',1,'']]],
-  ['diag_2eh_77',['Diag.h',['../Diag_8h.html',1,'']]],
-  ['displayer_78',['Displayer',['../classexeccommand__ns_1_1ExecCommand_1_1Displayer.html',1,'execcommand_ns::ExecCommand::Displayer'],['../classexeccommand__ns_1_1ExecCommand_1_1Displayer.html#a7aa30b72cba32d984c840662b0ebfdc5',1,'execcommand_ns::ExecCommand::Displayer::Displayer()']]],
-  ['dllexportflag_79',['DLLEXPORTFLAG',['../AnaCanScan_8cpp.html#a07e042970e18b919d409eb09524d625c',1,'DLLEXPORTFLAG():&#160;AnaCanScan.cpp'],['../STCanScan_8cpp.html#a07e042970e18b919d409eb09524d625c',1,'DLLEXPORTFLAG():&#160;STCanScan.cpp'],['../MockCanAccess_8cpp.html#a07e042970e18b919d409eb09524d625c',1,'DLLEXPORTFLAG():&#160;MockCanAccess.cpp']]],
-  ['dword_80',['DWORD',['../AnaCanScan_8h.html#ad342ac907eb044443153a22f964bf0af',1,'DWORD():&#160;AnaCanScan.h'],['../MockCanAccess_8h.html#ad342ac907eb044443153a22f964bf0af',1,'DWORD():&#160;MockCanAccess.h']]],
-  ['dynamicallyloadlib_81',['dynamicallyLoadLib',['../classCanModule_1_1CanLibLoader.html#a2443baffbe92cd107d09ea09e6cb7b28',1,'CanModule::CanLibLoader::dynamicallyLoadLib()'],['../classCanModule_1_1CanLibLoaderLin.html#aaef852fe6e43d093f58529fab0b2e841',1,'CanModule::CanLibLoaderLin::dynamicallyLoadLib()'],['../classCanModule_1_1CanLibLoaderWin.html#a8206b2ad7d352ecb8e119b55d2069bc0',1,'CanModule::CanLibLoaderWin::dynamicallyLoadLib()']]]
+  ['delete_5fmaps_87',['delete_maps',['../classCanModule_1_1Diag.html#aa79c20827b6e33fa35d4106d9aa479b8',1,'CanModule::Diag']]],
+  ['deletecanhandleofportip_88',['deleteCanHandleOfPortIp',['../classAnaCanScan.html#a1b62497ac5f0edcf9895c9f8cfad6b50',1,'AnaCanScan']]],
+  ['deviceid_89',['deviceID',['../structudevanalyserforpeak__ns_1_1UdevAnalyserForPeak_1_1PEAK__device__t.html#ac29ee824c7e4d8441a815af4c0523f6a',1,'udevanalyserforpeak_ns::UdevAnalyserForPeak::PEAK_device_t']]],
+  ['diag_90',['Diag',['../classCanModule_1_1Diag.html',1,'CanModule::Diag'],['../classCanModule_1_1Diag.html#a282a54b7181d085ba0abccca8fcb0874',1,'CanModule::Diag::Diag()']]],
+  ['diag_2ecpp_91',['Diag.cpp',['../Diag_8cpp.html',1,'']]],
+  ['diag_2eh_92',['Diag.h',['../Diag_8h.html',1,'']]],
+  ['displayer_93',['Displayer',['../classexeccommand__ns_1_1ExecCommand_1_1Displayer.html',1,'execcommand_ns::ExecCommand::Displayer'],['../classexeccommand__ns_1_1ExecCommand_1_1Displayer.html#a7aa30b72cba32d984c840662b0ebfdc5',1,'execcommand_ns::ExecCommand::Displayer::Displayer()']]],
+  ['dllexportflag_94',['DLLEXPORTFLAG',['../AnaCanScan_8cpp.html#a07e042970e18b919d409eb09524d625c',1,'DLLEXPORTFLAG():&#160;AnaCanScan.cpp'],['../STCanScan_8cpp.html#a07e042970e18b919d409eb09524d625c',1,'DLLEXPORTFLAG():&#160;STCanScan.cpp'],['../MockCanAccess_8cpp.html#a07e042970e18b919d409eb09524d625c',1,'DLLEXPORTFLAG():&#160;MockCanAccess.cpp']]],
+  ['dword_95',['DWORD',['../AnaCanScan_8h.html#ad342ac907eb044443153a22f964bf0af',1,'DWORD():&#160;AnaCanScan.h'],['../MockCanAccess_8h.html#ad342ac907eb044443153a22f964bf0af',1,'DWORD():&#160;MockCanAccess.h']]],
+  ['dynamicallyloadlib_96',['dynamicallyLoadLib',['../classCanModule_1_1CanLibLoader.html#a2443baffbe92cd107d09ea09e6cb7b28',1,'CanModule::CanLibLoader::dynamicallyLoadLib()'],['../classCanModule_1_1CanLibLoaderLin.html#aaef852fe6e43d093f58529fab0b2e841',1,'CanModule::CanLibLoaderLin::dynamicallyLoadLib()'],['../classCanModule_1_1CanLibLoaderWin.html#a8206b2ad7d352ecb8e119b55d2069bc0',1,'CanModule::CanLibLoaderWin::dynamicallyLoadLib()']]]
 ];