+++ Documentation/doxygen-html/search/classes_7.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['stcanscan_333',['STCanScan',['../classSTCanScan.html',1,'']]]
+  ['stcanscan_350',['STCanScan',['../classSTCanScan.html',1,'']]]
 ];