+++ Documentation/doxygen-html/search/defines_2.js
@@ -1,4 +1,9 @@
 var searchData=
 [
-  ['dllexportflag_633',['DLLEXPORTFLAG',['../AnaCanScan_8cpp.html#a07e042970e18b919d409eb09524d625c',1,'DLLEXPORTFLAG():&#160;AnaCanScan.cpp'],['../STCanScan_8cpp.html#a07e042970e18b919d409eb09524d625c',1,'DLLEXPORTFLAG():&#160;STCanScan.cpp'],['../MockCanAccess_8cpp.html#a07e042970e18b919d409eb09524d625c',1,'DLLEXPORTFLAG():&#160;MockCanAccess.cpp']]]
+  ['can_5fecho_5fmsg_662',['CAN_ECHO_MSG',['../MockCanAccess_8h.html#af8ac2352c89d82a9696dc9dc29ba8b77',1,'MockCanAccess.h']]],
+  ['canmodule_5fstatus_5fbp_5fanagate_663',['CANMODULE_STATUS_BP_ANAGATE',['../CCanAccess_8h.html#a1f1789533d0f909368bb51596939ef5b',1,'CCanAccess.h']]],
+  ['canmodule_5fstatus_5fbp_5fpeak_664',['CANMODULE_STATUS_BP_PEAK',['../CCanAccess_8h.html#a3c0a761bfb180463b0185b27edd53af5',1,'CCanAccess.h']]],
+  ['canmodule_5fstatus_5fbp_5fsock_665',['CANMODULE_STATUS_BP_SOCK',['../CCanAccess_8h.html#a6d0995692dcb6d1be30395a18d7683ba',1,'CCanAccess.h']]],
+  ['canmodule_5fstatus_5fbp_5fsystec_666',['CANMODULE_STATUS_BP_SYSTEC',['../CCanAccess_8h.html#af9d1d0a4353efec30879f97d7caebb76',1,'CCanAccess.h']]],
+  ['canmodule_5fversion_667',['CanModule_VERSION',['../CCanAccess_8h.html#ab5f1e478bce31c034108ce1e0113a7d8',1,'VERSION.h']]]
 ];