+++ Documentation/doxygen-html/search/enumvalues_3.js
@@ -1,5 +1,4 @@
 var searchData=
 [
-  ['sendfail_622',['sendFail',['../namespaceCanModule.html#a2f4fee79cc7ba53467c2538fe6b9adbbae323a25a871808f9a0df9a72ce954585',1,'CanModule']]],
-  ['singlebus_623',['singleBus',['../namespaceCanModule.html#a2062089fa4fa87c8354295f2a8fe31a5a5f768fefafa6db21eee26885afb6163d',1,'CanModule']]]
+  ['timeoutonreception_646',['timeoutOnReception',['../namespaceCanModule.html#a2f4fee79cc7ba53467c2538fe6b9adbbaba350cf15e5a1471109f17351afb401a',1,'CanModule']]]
 ];