+++ Documentation/doxygen-html/search/functions_1.js
@@ -1,5 +1,6 @@
 var searchData=
 [
-  ['addcanhandleofportip_370',['addCanHandleOfPortIp',['../classAnaCanScan.html#a3e2635663ebdfac68591d6a0577fa951',1,'AnaCanScan']]],
-  ['anacanscan_371',['AnaCanScan',['../classAnaCanScan.html#af6c27ae08034c5da64a54de29e2e875e',1,'AnaCanScan::AnaCanScan()'],['../classAnaCanScan.html#af05ec5dfd6ac56764c6db8e721648b75',1,'AnaCanScan::AnaCanScan(AnaCanScan const &amp;other)=delete']]]
+  ['addcanhandleofportip_388',['addCanHandleOfPortIp',['../classAnaCanScan.html#a3e2635663ebdfac68591d6a0577fa951',1,'AnaCanScan']]],
+  ['ana_5fcangeterrortext_389',['ana_canGetErrorText',['../classAnaCanScan.html#a1425dfc20e08b545fcdf49415ac15ec9',1,'AnaCanScan']]],
+  ['anacanscan_390',['AnaCanScan',['../classAnaCanScan.html#af6c27ae08034c5da64a54de29e2e875e',1,'AnaCanScan::AnaCanScan()'],['../classAnaCanScan.html#af05ec5dfd6ac56764c6db8e721648b75',1,'AnaCanScan::AnaCanScan(AnaCanScan const &amp;other)=delete']]]
 ];