+++ Documentation/doxygen-html/search/functions_10.js
@@ -1,11 +1,13 @@
 var searchData=
 [
-  ['timesinceopened_487',['timeSinceOpened',['../classCanModule_1_1CanStatistics.html#a08d5c282de626e89ebf508291d1bde33',1,'CanModule::CanStatistics']]],
-  ['timesincereceived_488',['timeSinceReceived',['../classCanModule_1_1CanStatistics.html#a44da075145ae6dd716bcb9432e254ce7',1,'CanModule::CanStatistics']]],
-  ['timesincetransmitted_489',['timeSinceTransmitted',['../classCanModule_1_1CanStatistics.html#a82bf9c9ca0e1578dbd962ee3d2e05fa6',1,'CanModule::CanStatistics']]],
-  ['tohexstring_490',['toHexString',['../classCanModule_1_1CanModuleUtils.html#aeb2a2ee933f230f2c82cca7e45c7b6ad',1,'CanModule::CanModuleUtils']]],
-  ['tostring_491',['toString',['../classCanModule_1_1CanModuleUtils.html#a535734490dee7c5c0ff0f51cb3875289',1,'CanModule::CanModuleUtils']]],
-  ['totalreceived_492',['totalReceived',['../classCanModule_1_1CanStatistics.html#a5a06575683a46d25b4a1cc7d6859938b',1,'CanModule::CanStatistics']]],
-  ['totaltransmitted_493',['totalTransmitted',['../classCanModule_1_1CanStatistics.html#a62ee3d5dd02a649066747cb2e2339b9d',1,'CanModule::CanStatistics']]],
-  ['txrate_494',['txRate',['../classCanModule_1_1CanStatistics.html#ad8a95eeeea9fe50c9fe5465b3bde95d9',1,'CanModule::CanStatistics']]]
+  ['testthread_507',['testThread',['../SockCanScan_8cpp.html#ab85356f782f86b4f9882ab007a4363c4',1,'SockCanScan.cpp']]],
+  ['testthread2_508',['testThread2',['../SockCanScan_8cpp.html#a1ff03dbfb54bd09b1d1d521060639cc7',1,'SockCanScan.cpp']]],
+  ['timesinceopened_509',['timeSinceOpened',['../classCanModule_1_1CanStatistics.html#a08d5c282de626e89ebf508291d1bde33',1,'CanModule::CanStatistics']]],
+  ['timesincereceived_510',['timeSinceReceived',['../classCanModule_1_1CanStatistics.html#a44da075145ae6dd716bcb9432e254ce7',1,'CanModule::CanStatistics']]],
+  ['timesincetransmitted_511',['timeSinceTransmitted',['../classCanModule_1_1CanStatistics.html#a82bf9c9ca0e1578dbd962ee3d2e05fa6',1,'CanModule::CanStatistics']]],
+  ['tohexstring_512',['toHexString',['../classCanModule_1_1CanModuleUtils.html#aeb2a2ee933f230f2c82cca7e45c7b6ad',1,'CanModule::CanModuleUtils']]],
+  ['tostring_513',['toString',['../classCanModule_1_1CanModuleUtils.html#a535734490dee7c5c0ff0f51cb3875289',1,'CanModule::CanModuleUtils']]],
+  ['totalreceived_514',['totalReceived',['../classCanModule_1_1CanStatistics.html#a5a06575683a46d25b4a1cc7d6859938b',1,'CanModule::CanStatistics']]],
+  ['totaltransmitted_515',['totalTransmitted',['../classCanModule_1_1CanStatistics.html#a62ee3d5dd02a649066747cb2e2339b9d',1,'CanModule::CanStatistics']]],
+  ['txrate_516',['txRate',['../classCanModule_1_1CanStatistics.html#ad8a95eeeea9fe50c9fe5465b3bde95d9',1,'CanModule::CanStatistics']]]
 ];