+++ Documentation/doxygen-html/search/functions_12.js
@@ -1,14 +1,4 @@
 var searchData=
 [
-  ['_7eanacanscan_498',['~AnaCanScan',['../classAnaCanScan.html#a2c8256b5096aa1351fbd9f3f2f717180',1,'AnaCanScan']]],
-  ['_7ecanlibloader_499',['~CanLibLoader',['../classCanModule_1_1CanLibLoader.html#aaf19503282f3fe164c434674721c4972',1,'CanModule::CanLibLoader']]],
-  ['_7ecanlibloaderlin_500',['~CanLibLoaderLin',['../classCanModule_1_1CanLibLoaderLin.html#a91636138f96d42a3a835c2a38bae3ec6',1,'CanModule::CanLibLoaderLin']]],
-  ['_7ecanlibloaderwin_501',['~CanLibLoaderWin',['../classCanModule_1_1CanLibLoaderWin.html#a39da78f1fde4fef2b6b17107e6767962',1,'CanModule::CanLibLoaderWin']]],
-  ['_7eccanaccess_502',['~CCanAccess',['../classCanModule_1_1CCanAccess.html#ac96f1cce6beda4309cdfb6651d3ce547',1,'CanModule::CCanAccess']]],
-  ['_7ecsockcanscan_503',['~CSockCanScan',['../classCSockCanScan.html#a243f91e993fcc96ddd5308c1425a130c',1,'CSockCanScan']]],
-  ['_7ediag_504',['~Diag',['../classCanModule_1_1Diag.html#a2725bd2ef8f986c94a0f025ac6b95240',1,'CanModule::Diag']]],
-  ['_7emockcanaccess_505',['~MockCanAccess',['../classMockCanAccess.html#a87261ff4a42e3b4d0f02939e7adc39d3',1,'MockCanAccess']]],
-  ['_7epkcanscan_506',['~PKCanScan',['../classPKCanScan.html#a924778bdca171eb7f3f597dfe9f99fb4',1,'PKCanScan']]],
-  ['_7estcanscan_507',['~STCanScan',['../classSTCanScan.html#a2f3398dc992ebd90002c179305c55d54',1,'STCanScan']]],
-  ['_7eudevanalyserforpeak_508',['~UdevAnalyserForPeak',['../classudevanalyserforpeak__ns_1_1UdevAnalyserForPeak.html#a9120a53396820347617a1a4eba992a97',1,'udevanalyserforpeak_ns::UdevAnalyserForPeak']]]
+  ['vendorbaudrate2userbaudrate_520',['vendorBaudRate2UserBaudRate',['../classSTCanScan.html#aa57c48594c8a70a0331f07fb4ad208fd',1,'STCanScan']]]
 ];